requirements.txt
# python=3.11
python-dotenv==1.0.1
click==8.1.7
pyyaml==6.0.1
tqdm==4.66.4
numpy==1.26.4
matplotlib==3.8.4
pandas==2.2.2
scikit-learn==1.4.2
nltk==3.8.1
spacy==3.7.4

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.0+cu118

transformers==4.40.1
huggingface-hub==0.23.0